Update README.md
Task 1: Created a step-by-step, detailed description of how to use Webstorm for a first-timer. Includes downloading instruction for Webstorm and Git, how to create and edit an HTML file, and test it. Goes over detailed instruction on how to start uploading files on Github, creating repositories, and locating branches/repositories. 
Task 2: Defined the Github commands in a language where it would be easy for non-technical or those with less experience in Github/coding will be able to understand. Bolded each word to separate it from the sentences.

Edited the README.md file and added these new changes.